pyproject.toml
[project]
name = "esquite"
version = "0.1.0"
description = "Framework para corpus paralelos | Framework for parallel corpora"
authors = [
    { name = "Diego Barriga", email = "dbarriga@ciencias.unam.mx" },
    { name = "Comunidad Elotl", email = "contacto@elotl.mx" },
]
requires-python = ">=3.12"
license = "MPL-2.0"
dependencies = [
    "Django>=4.2,<5",
    "djangorestframework>=3.13.1,<4",
    "elasticsearch>=7.16.3,<8",
    "GitPython>=3.1.26,<4",
    "requests>=2.27.1,<3",
    "click>=8.0.3,<9",
    "PyYAML~=6.0",
    "drf-yasg>=1.20.0,<2",
    "bs4>=0.0.2",
]

[dependency-groups]
dev = [
    "Sphinx>=4.4.0,<5",
    "sphinxcontrib-redoc>=1.6.0,<2",
]

# TODO: Think if we need a build system
